+++ src/medication-schedules/medication-schedules.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
 import { MedicationSchedulesController } from './medication-schedules.controller';
 import { MedicationSchedulesService } from './medication-schedules.service';
-import { MedicationSchedule } from 'src/app.entities';
+
 import { MikroOrmModule } from '@mikro-orm/nestjs';
+import { MedicationSchedule } from 'src/entities/medication-schedule.entity';
 
 @Module({
   imports: [MikroOrmModule.forFeature([MedicationSchedule])],
