+++ src/mikro-orm.config.ts
@@ -11,8 +11,8 @@ const config: Options = {
   password: process.env.DB_PASSWORD,
   dbName: process.env.DB_NAME,
   driver: MySqlDriver,
-  entities: ['./dist/*.entities.js'],
-  entitiesTs: ['./src/*.entities.ts'],
+  entities: ['./dist/entities/*.entity.js'],
+  entitiesTs: ['./src/entities/*.entity.ts'],
   metadataProvider: TsMorphMetadataProvider,
   debug: true,
   extensions: [Migrator],
