+++ src/user/users.module.ts
@@ -1,9 +1,9 @@
 import { MikroOrmModule } from '@mikro-orm/nestjs';
 import { Module } from '@nestjs/common';
-import { User } from 'src/app.entities';
 import { UsersService } from './users.service';
 import { UserController } from './users.controller';
 import { GraphQLClientService } from 'src/utils/graphql/graphql.service';
+import { User } from 'src/entities/user.entity';
 
 @Module({
   imports: [MikroOrmModule.forFeature([User])],
